add ESlint support for monorepo
- install eslint and
 with cmd: `lerna add eslint packages/* --dev && lerna add eslint-config-peggy packages/* --dev`
as this article mentioned:
https://medium.com/rewrite-tech/how-to-create-a-monorepo-with-lerna-3ed6dfec5021
Add ESLint to all packages
why to install these dependecies in the package folder and not at the root folder?
You want to make the packages independent from the outside of its folder as much as possible.
In that way just put all the packages into the package folder as long as it is not depending on tools outside of the monorepo

- add eslintrc.js and eslintignore files into each package folder, so they are independent and can be customized later if needed
- add source.fixAll.eslint": true in settings.json with Prettier in root folder, so we can auto fix the formatting issues
- add lint script cmd in each packages' package.json file and add 2 root level linting cmds in root folder,
  so we can run `npm run lint` for all packages parallel